+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [1.1.0] - 20-05-2026
+
+### Added
+- Added support for selecting mail-enabled security groups
+
+### Changed
+- Repository structure aligned with the Create and Update repository
+
 ## [1.0.0] - 18-02-2026
 
 Initial release